Split out computeLayout module
